backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.28.1
json-repair==0.46.3
